Exploring the Virtual Realm : METAVERSE
The VR open auditorium in the metaverse is a virtual space where people gather to attend events, presentations, and performances in a shared digital environment. Utilizing VR technology, participants can immerse themselves in a variety of experiences, from virtual conferences and interview to educational seminars and art exhibitions. This digital auditorium offers the freedom of movement, social interaction, and access to diverse content, bridging geographical barriers and fostering community engagement in a virtual realm.